Added Blended PET and potential melt algorithms, support for Borland C compiler [BENCHMARKED]
Added Blended PET and Blended Potential melt algorithms
-new algorithms declared (RavenInclude.h)
-new members and specification routines (SetPETBlendValues,SetPotMeltBlendValue) in Model.cpp/.h
-Parsing support, new ParsePETMethod, ParsePotMeltMethod routines (ParseInput.cpp)
-implementation (Evaporation.cpp, PotentialMelt.cpp)

Borland C compiler warnings (from Pierre @CMM)
-declaration of constants as size_t (CommonFunctions.cpp)
-unused variables found (CustomOutput.cpp; DDS.cpp;Diagnostics.cpp;LatFlush.cpp;ModelForcingGrids.cpp; 
REservoir.cpp;PotentialMetl.cpp;PrairieSnow.cpp;SnowBalance.cpp;SoilClass.cpp;SoilEvaporation.cpp;StandardOutput.cpp;
SubBasin.cpp;UpdateForcings.cpp;VegetationClass.cpp)
-if (true) statements in ParsePropertyFile.cpp
-removed breaks from return-filled case statements (Radiation.cpp)
-moved member declaration within diversion struct (SubBasin.h)
-changed handling of CTimeSeriesABC arg list (TimeSeries.cpp, TimeSeriesABC.h; TimeSeriesABC.cpp)

Support for Basin INFLOW override in ParseFEWSStateFile (ParseFEWSRunInfo.cpp)

Fixed such that forcings.csv and .nc aren't both generated (StandardOutput.cpp)

Fixed bugin ReservoirStages.nc for FEWS work (StandardOutput.cpp)

Clean up
-removed unused CModel::_NumSnowLayers, CModel::_NumAquiferLayers members (Model.cpp/.h;ModelInitialize.cpp)
-removed CModelABC::GetNumAquiferLayers() (ModelABC.h)


git-svn-id: https://chyms.nrc.gc.ca/svn/raven/trunk@343 9a3899ca-7adb-4d27-9b38-7f4ebb3123b0

+++ src/CommonFunctions.cpp
@@ -404,11 +404,11 @@ time_struct DateStringToTimeStruct(const string sDate, string sTime, const int c
 {
 
   static time_struct tt;
-  if (sDate.length()!=10){
+  if (sDate.length()!=(size_t)(10)){
     string errString = "DateStringToTimeStruct: Invalid date format used: "+sDate;
     ExitGracefully(errString.c_str(),BAD_DATA);
   }
-  if (sTime.length()<7)  {
+  if (sTime.length()<(size_t)(7))  {
     string errString = "DateStringToTimeStruct: Invalid time format used (hourstamp): "+sTime;
     ExitGracefully(errString.c_str(),BAD_DATA);
   }
@@ -513,7 +513,7 @@ time_struct TimeStructFromNetCDFString(const string unit_t_str,const string time
     if(!strstr(tmp.substr(start+16,1).c_str(),":")) { tmp.insert(start+14,"0"); } // second colon in time - fixes 11:0:00 (?)
   }
   else {
-    if(strlen(tmp.c_str())==start+9) { tmp.insert(start+8,"0"); } // second dash in date - fixes YYYY-MM-d 
+    if(strlen(tmp.c_str())==(size_t)(start+9)) { tmp.insert(start+8,"0"); } // second dash in date - fixes YYYY-MM-d 
   }
   string sTime,sDate;
   sDate = tmp.substr(start,10); //YYYY-MM-DD
@@ -522,7 +522,7 @@ time_struct TimeStructFromNetCDFString(const string unit_t_str,const string time
   //cout<<"sTime"<<sTime<<" sDate"<< sDate<<" "<<unit_t_str<<" "<<date_only<<" "<<tmp<<endl;
 
   timezone=0;
-  if((strlen(tmp.c_str())-start)==26) {
+  if((strlen(tmp.c_str())-start)==(size_t)(26)) {
     if(!strcmp(tmp.substr(start+22,1).c_str(),"+")) {
       timezone=(double)(s_to_i(tmp.substr(start+23,4).c_str()))/HR_PER_DAY/100; //time zone, in days
     }
